subgraphs/shibadrop-erc1155/tests/erc-1155-shiba-drop-clone-factory.test.ts
import {
  assert,
  describe,
  test,
  clearStore,
  beforeAll,
  afterAll
} from "matchstick-as/assembly/index"
import { Address } from "@graphprotocol/graph-ts"
import { NewInstance1155 } from "../generated/schema"
import { NewInstance1155 as NewInstance1155Event } from "../generated/ERC1155ShibaDropCloneFactory/ERC1155ShibaDropCloneFactory"
import { handleNewInstance1155 } from "../src/erc-1155-shiba-drop-clone-factory"
import { createNewInstance1155Event } from "./erc-1155-shiba-drop-clone-factory-utils"

// Tests structure (matchstick-as >=0.5.0)
// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0

describe("Describe entity assertions", () => {
  beforeAll(() => {
    let instance = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let creator = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let newNewInstance1155Event = createNewInstance1155Event(instance, creator)
    handleNewInstance1155(newNewInstance1155Event)
  })

  afterAll(() => {
    clearStore()
  })

  // For more test scenarios, see:
  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test

  test("NewInstance1155 created and stored", () => {
    assert.entityCount("NewInstance1155", 1)

    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
    assert.fieldEquals(
      "NewInstance1155",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "instance",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "NewInstance1155",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "creator",
      "0x0000000000000000000000000000000000000001"
    )

    // More assert options:
    // https://thegraph.com/docs/en/developer/matchstick/#asserts
  })
})
